web_app/requirements.txt
flask==2.3.3
pillow==10.0.0
werkzeug==2.3.7
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7 